Refactor: Remove UIOptionsSelector component and implement PokemonFilters
- Deleted UIOptionsSelector component and its associated styles.
- Added new PokemonFilters component with comprehensive filtering options for Pokémon.
- Introduced new SCSS styles for PokemonFilters to enhance UI.
- Updated imports across various components to reflect new file structure and enums.
- Refactored layout and sprite-related hooks to use new model paths.
- Adjusted TypeScript configurations to align with new model structure.
