--- src/components/CardBackgroundSelector.scss
@@ -1,41 +1,0 @@
-// Importar el sistema de temas
-@use '@/assets/styles' as *;
-
-.card-background-selector {
-	display: flex;
-	align-items: center;
-	gap: 0.75rem;
-	margin-bottom: 1rem;
-	padding: 0.75rem;
-	background: var(--overlay-light);
-	border-radius: 8px;
-	border: 1px solid var(--border-overlay-light);
-
-	.selector-label {
-		font-size: 0.9rem;
-		color: var(--text-primary);
-		font-weight: 500;
-		white-space: nowrap;
-	}
-
-	.background-selector {
-		padding: 0.5rem 0.75rem;
-		border: 1px solid var(--border-overlay-medium);
-		border-radius: 6px;
-		background: var(--overlay-light);
-		color: var(--text-primary);
-		font-size: 0.9rem;
-		min-width: 180px;
-		cursor: pointer;
-
-		&:focus {
-			outline: 2px solid var(--color-focus-ring);
-			outline-offset: 2px;
-		}
-
-		option {
-			background: var(--bg-secondary);
-			color: var(--text-primary);
-		}
-	}
-}
