--- src/components/Header.scss
@@ -1,100 +1,0 @@
-// Header.scss - Cabecera principal de la aplicación
-@import '../App.scss';
-
-.app-header {
-	background: var(--overlay-light);
-	border-bottom: 1px solid var(--border-overlay-light);
-	padding: 0;
-	position: sticky;
-	top: 0;
-	z-index: 100;
-	backdrop-filter: blur(10px);
-
-	.header-content {
-		max-width: 1200px;
-		margin: 0 auto;
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		padding: 12px 20px;
-	}
-
-	.header-logo {
-		display: flex;
-		align-items: center;
-		text-decoration: none;
-		transition: all 0.2s ease;
-
-		&:hover {
-			transform: translateY(-1px);
-		}
-
-		.logo-image {
-			height: 40px;
-			width: auto;
-		}
-	}
-
-	.header-nav {
-		display: flex;
-		gap: 16px;
-		align-items: center;
-	}
-
-	.nav-link {
-		display: flex;
-		align-items: center;
-		gap: 6px;
-		padding: 8px 16px;
-		text-decoration: none;
-		color: var(--text-secondary);
-		font-weight: 500;
-		font-size: 0.9rem;
-		border-radius: 8px;
-		border: 1px solid transparent;
-		transition: all 0.2s ease;
-
-		&:hover {
-			background: var(--overlay-medium);
-			color: var(--text-primary);
-			border-color: var(--border-overlay-medium);
-			transform: translateY(-1px);
-		}
-
-		&.active {
-			background: var(--color-primary-alpha-20);
-			color: var(--color-primary);
-			border-color: var(--color-primary-alpha-50);
-			font-weight: 600;
-		}
-	}
-
-	// Responsive design
-	@media (max-width: 768px) {
-		.header-content {
-			padding: 8px 16px;
-		}
-
-		.header-logo .logo-image {
-			height: 32px;
-		}
-
-		.nav-link {
-			padding: 6px 12px;
-			font-size: 0.85rem;
-		}
-	}
-
-	@media (max-width: 480px) {
-		.header-content {
-			flex-direction: column;
-			gap: 12px;
-			padding: 12px 16px;
-		}
-
-		.header-nav {
-			width: 100%;
-			justify-content: center;
-		}
-	}
-}
