+++ src/components/Home/components/HomeComponent.tsx
@@ -3,7 +3,7 @@ import { downloadFileById, downloadPkmFileFromDisk } from '@/services/Pokemon'
 import type { PokemonListFilterDto } from '@/models/Pokemon'
 import type { PokemonListItemDto, TagDto } from '@/models/api/types'
 import { ConfirmDialog } from '@/ConfirmDialog'
-import { PokemonFilters, PokemonCard, PokemonListRow } from '@/components'
+import { PokemonFilters, PokemonCard, PokemonListRow } from '@/components/elements'
 import { TagManager } from '@/components/elements/TagManager/TagManager'
 import { useViewMode } from '@/hooks/useViewMode'
 import { useSpriteType } from '@/hooks/useSpriteType'
