+++ src/components/Settings/components/Settings.scss
@@ -146,3 +146,66 @@
 		padding: 16px;
 	}
 }
+
+.ui-options-selector {
+	background: var(--overlay-light);
+	border: 1px solid var(--border-overlay-light);
+	border-radius: 8px;
+	padding: 12px;
+	margin-bottom: 16px;
+
+	.selector-title {
+		color: var(--text-primary);
+		font-size: 0.9rem;
+		font-weight: 600;
+		margin: 0 0 12px 0;
+		display: flex;
+		align-items: center;
+		gap: 8px;
+		padding: 12px 0 0 12px;
+	}
+
+	.options-flex {
+		display: flex;
+		flex-direction: row;
+		gap: 16px;
+
+		.option-group {
+			border: 1px solid var(--border-overlay-light);
+			border-radius: 8px;
+			padding: 12px;
+
+			.option-label {
+				color: var(--text-primary);
+				font-size: 0.9rem;
+				font-weight: 600;
+				margin: 0 0 8px 0;
+				display: flex;
+				align-items: center;
+				gap: 8px;
+			}
+
+			.option-select {
+				width: 100%;
+				padding: 0.5rem 0.75rem;
+				border: 1px solid var(--border-overlay-medium);
+				border-radius: 6px;
+				background: var(--overlay-light);
+				color: var(--text-primary);
+				font-size: 0.9rem;
+				cursor: pointer;
+				margin-bottom: 12px;
+
+				&:focus {
+					outline: 2px solid var(--color-focus-ring);
+					outline-offset: 2px;
+				}
+
+				option {
+					background: var(--bg-secondary);
+					color: var(--text-primary);
+				}
+			}
+		}
+	}
+}
