+++ src/components/elements/PokemonCard/PokemonCard.tsx
@@ -2,10 +2,10 @@ import type { PokemonListItemDto } from '../../../models/api/types'
 import { usePokeBallIcon } from '../../../hooks/useCachedAssets'
 import { usePokemonInfo } from '../../../hooks/usePokemonInfo'
 import { getTypeIconUrl } from '../../../utils'
-import { getTypeNameFromId } from '../../../enums/PokemonTypes'
-import { getBallNameFromId } from '../../../enums/PokemonBalls'
+import { getTypeNameFromId } from '../../../models/enums/PokemonTypes'
+import { getBallNameFromId } from '../../../models/enums/PokemonBalls'
 import { useCardBackgroundType } from '../../../hooks/useCardBackgroundType'
-import { CardBackgroundType } from '../../../enums/CardBackgroundTypes'
+import { CardBackgroundType } from '../../../models/enums/CardBackgroundTypes'
 import { getComputedTypeColor } from '../../../utils/typeColors'
 import './PokemonCard.scss'
