+++ src/components/elements/index.ts
@@ -1,3 +1,8 @@
-export { ThemeSelector } from './ThemeSelector/ThemeSelector'
-export { TagManager } from './TagManager/TagManager'
 export { LayoutSelector } from './LayoutSelector/LayoutSelector'
+export { Modal } from './Modal/Modal'
+export { PokemonCard } from './PokemonCard/PokemonCard'
+export { PokemonFilters } from './Filters/PokemonFilters'
+export { PokemonListRow } from './PokemonRow/PokemonListRow'
+export { TagManager } from './TagManager/TagManager'
+export { ThemeSelector } from './ThemeSelector/ThemeSelector'
+export { UploadAndScanFiles } from './UploadAndScanFiles/UploadAndScanFiles'
