--- src/components/index.ts
@@ -1,5 +1,0 @@
-// Re-export the new modular filters system
-export { PokemonFilters } from './PokemonFilters'
-export { PokemonCard } from './elements/PokemonCard/PokemonCard'
-export { PokemonListRow } from './elements/PokemonRow/PokemonListRow'
-export { PokemonCacheDiagnostics } from './PokemonCacheDiagnostics'
