+++ src/hooks/useCardBackgroundType.ts
@@ -1,6 +1,6 @@
 import { useAppSelector, useAppDispatch } from '@/store/hooks'
 import { setBackgroundType as setBackgroundTypeAction } from '@/store/features/styleSettings'
-import type { CardBackgroundTypeName } from '../enums/CardBackgroundTypes'
+import type { CardBackgroundTypeName } from '../models/enums/CardBackgroundTypes'
 
 export function useCardBackgroundType() {
 	const backgroundType = useAppSelector((state) => state.styleSettings.backgroundType)
