+++ src/hooks/useLayout.ts
@@ -1,5 +1,6 @@
 import { useAppSelector, useAppDispatch } from '@/store/hooks'
-import { setLayoutType, type LayoutType } from '@/store/features/layout'
+import { setLayoutType } from '@/store/features/layout'
+import type { LayoutType } from '@/models/store/Layout'
 
 export function useLayout() {
 	const layoutType = useAppSelector((state) => state.layout.layoutType)
