+++ src/models/enums/CardBackgroundTypes.ts
