+++ src/models/enums/SpriteTypes.ts
